Move logging related classes to Dynamo.Logging
