+++ src/DynamoCore/DynamoCore.csproj
@@ -217,7 +217,7 @@ limitations under the License.
     <Compile Include="Nodes\FunctionCallBase.cs" />
     <Compile Include="Nodes\ZeroTouch\UnresolvedFunctionException.cs" />
     <Compile Include="Search\BrowserItem.cs" />
-    <Compile Include="Services\StabilityTracking.cs" />
+    <Compile Include="Logging\StabilityTracking.cs" />
     <Compile Include="Engine\AstBuilder.cs" />
     <Compile Include="Library\FunctionType.cs" />
     <Compile Include="Engine\EngineController.cs" />
@@ -233,9 +233,9 @@ limitations under the License.
     <Compile Include="Core\PreferenceSettings.cs" />
     <Compile Include="Core\DynamoLogger.cs" />
     <Compile Include="Core\DynamoSelection.cs" />
-    <Compile Include="Services\Heartbeat.cs" />
-    <Compile Include="Services\Log.cs" />
-    <Compile Include="Services\Logger.cs" />
+    <Compile Include="Logging\Heartbeat.cs" />
+    <Compile Include="Logging\Log.cs" />
+    <Compile Include="Logging\Logger.cs" />
     <Compile Include="Models\DynamoModel.cs" />
     <Compile Include="Models\ModelBase.cs" />
     <Compile Include="Models\NoteModel.cs" />
