+++ src/DynamoCore/Logging/Log.cs
@@ -6,13 +6,13 @@
  */
 
 using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Net;
 using System.Text;
 using System.Threading;
-using System.Net;
-using System.IO;
-using System.Collections.Generic;
 
-namespace net.riversofdata.dhlogger
+namespace Dynamo.Logging
 {
     internal class Log : IDisposable
     {
