+++ src/DynamoCore/Logging/LogSource.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using Dynamo.Interfaces;
 
 namespace Dynamo.Logging
