+++ src/DynamoCoreWpf/ViewModels/Search/SearchViewModel.cs
@@ -7,6 +7,7 @@ using System.Text.RegularExpressions;
 using System.Windows;
 using System.Windows.Media;
 using Dynamo.Interfaces;
+using Dynamo.Logging;
 using Dynamo.Nodes;
 using Dynamo.Search;
 using Dynamo.Search.SearchElements;
